move admin to it's own package
