+++ dgraph/cmd/graphql/admin/admin.go
@@ -14,21 +14,15 @@
  * limitations under the License.
  */
 
-package graphql
+package admin
 
 import (
-	"context"
-	"fmt"
 	"time"
 
 	"github.com/dgraph-io/dgo/v2"
-	dgoapi "github.com/dgraph-io/dgo/v2/protos/api"
 	"github.com/dgraph-io/dgraph/dgraph/cmd/graphql/resolve"
 	"github.com/dgraph-io/dgraph/dgraph/cmd/graphql/schema"
 	"github.com/dgraph-io/dgraph/dgraph/cmd/graphql/web"
-	"github.com/dgraph-io/dgraph/gql"
-	"github.com/golang/glog"
-	"github.com/pkg/errors"
 )
 
 const (
@@ -116,57 +110,11 @@ const (
 	addSchema(input: SchemaInput!) : AddSchemaPayload
  }
  `
-
-	// Heath status codes
-	errNoConnection healthStatus = "ErrNoConnection"
-	noGraphQLSchema healthStatus = "noGraphQLSchema"
-	healthy         healthStatus = "Healthy"
 )
-
-type healthStatus string
 
 type schemaDef struct {
 	Schema string    `json:"schema,omitempty"`
 	Date   time.Time `json:"date,omitempty"`
-}
-
-type healthResolver struct {
-	status healthStatus
-}
-
-// A addSchemaResolver serves as the mutation rewriter and executor in handling
-// the addSchema mutation.
-type addSchemaResolver struct {
-	// the Dgraph that gets its schema changed
-	dgraph *dgo.Dgraph
-
-	// schema that is generated from the mutation input
-	newGQLSchema    schema.Schema
-	newDgraphSchema string
-
-	// The underlying executor and rewriter that persist the schema into Dgraph as
-	// GraphQL metadata
-	baseMutationRewriter resolve.MutationRewriter
-	baseMutationExecutor resolve.MutationExecutor
-
-	// The GraphQL server that's being admin'd
-	gqlServer web.IServeGraphQL
-
-	// When the schema changes, we use these to create a new RequestResolver for
-	// the main graphql endpoint (gqlServer) and thus refresh the API.
-	fns               *resolve.ResolverFns
-	withIntrospection bool
-}
-
-var statusMessage = map[healthStatus]string{
-	errNoConnection: "Unable to contact Dgraph",
-	noGraphQLSchema: "Dgraph connection established but there's no GraphQL schema.",
-	healthy:         "Dgraph connection established and serving GraphQL schema.",
-}
-
-func (hr *healthResolver) Query(ctx context.Context, query *gql.GraphQuery) ([]byte, error) {
-	s := hr.status
-	return []byte(fmt.Sprintf(`{"message":"%s","status":%s`, statusMessage[s], string(s))), nil
 }
 
 // NewAdminResolver creates a GraphQL request resolver for the /admin endpoint.
@@ -226,76 +174,4 @@ func newAdminResolverFactory(
 					resolve.StdQueryCompletion())
 			}).
 		WithSchemaIntrospection()
-}
-
-func (asr *addSchemaResolver) Rewrite(m schema.Mutation) (*gql.GraphQuery, []*dgoapi.Mutation, error) {
-	sch, err := getSchemaInput(m)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	schHandler, err := schema.NewHandler(sch)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	asr.newGQLSchema, err = schema.FromString(schHandler.GQLSchema())
-	if err != nil {
-		return nil, nil, err
-	}
-
-	asr.newDgraphSchema = schHandler.DGSchema()
-
-	m.SetArgTo(schema.InputArgName, map[string]interface{}{"schema": sch, "date": time.Now()})
-	return asr.baseMutationRewriter.Rewrite(m)
-}
-
-func (asr *addSchemaResolver) FromMutationResult(
-	mutation schema.Mutation,
-	assigned map[string]string,
-	mutated map[string][]string) (*gql.GraphQuery, error) {
-	return asr.baseMutationRewriter.FromMutationResult(mutation, assigned, mutated)
-}
-
-func (asr *addSchemaResolver) Mutate(
-	ctx context.Context,
-	query *gql.GraphQuery,
-	mutations []*dgoapi.Mutation) (map[string]string, map[string][]string, error) {
-
-	assigned, mutated, err := asr.baseMutationExecutor.Mutate(ctx, query, mutations)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	if glog.V(3) {
-		glog.Infof("Altering Dgraph schema:\n\n%s\n", asr.newDgraphSchema)
-	}
-
-	err = asr.dgraph.Alter(ctx, &dgoapi.Operation{Schema: asr.newDgraphSchema})
-	if err != nil {
-		return nil, nil, schema.GQLWrapf(err,
-			"succeeded in saving GraphQL schema but failed to alter Dgraph schema "+
-				"(you should retry)")
-	}
-
-	resolverFactory := resolve.NewResolverFactory().
-		WithConventionResolvers(asr.newGQLSchema, asr.fns)
-
-	if asr.withIntrospection {
-		resolverFactory.WithSchemaIntrospection()
-	}
-
-	resolvers := resolve.New(asr.newGQLSchema, resolverFactory)
-	asr.gqlServer.ServeGQL(resolvers)
-
-	return assigned, mutated, nil
-}
-
-func getSchemaInput(m schema.Mutation) (string, error) {
-	input, ok := m.ArgValue(schema.InputArgName).(map[string]interface{})
-	if !ok {
-		return "", errors.New("couldn't get input argument")
-	}
-
-	return input["schema"].(string), nil
 }
